Added constructor of class Game

+++ src/Model/Asteroid.java
@@ -1,17 +1,17 @@
 package Model;
 
 public class Asteroid {
-    private double pos_x, pos_y;
+    private int pos_x, pos_y;
 
     public Asteroid(int pos_x, int pos_y){
         this.pos_x = pos_x;
         this.pos_y = pos_y;
     }
 
-    public double get_pos_x(){
+    public int get_pos_x(){
         return pos_x;
     }
-    public double get_pos_y(){
+    public int get_pos_y(){
         return pos_y;
     }
